consolidate instances of `loadInst`, so code isn't repeated
